requirements.txt
Django<=1.10.1
psycopg2<=2.6.2
django-dotenv<=1.4.1
django-getenv<=1.3.1
djangorestframework<=3.4.6
djangorestframework-gis<=0.10.1
django-geojson<=2.9.1
django-cors-headers
requests
python-dateutil
plotly
